+++ core/calendar/src/main/java/com/cheeruplife/core/calendar/model/CalendarType.kt
@@ -1,6 +1,6 @@
 package com.cheeruplife.core.calendar.model
 
-enum class CalendarType {
+internal enum class CalendarType {
     SMALL_CALENDAR,
     NORMAL_CALENDAR,
     EXPANDED_CALENDAR,
