+++ core/calendar/src/main/java/com/cheeruplife/core/calendar/model/ScrollState.kt
@@ -1,6 +1,6 @@
 package com.cheeruplife.core.calendar.model
 
-enum class ScrollState {
+internal enum class ScrollState {
     UP_SCROLL,
     DOWN_SCROLL,
     NONE,
